// src/views/collection/Seo/__generated__/CollectionSeoFragment_storeCollection.graphql.ts
/**
 * Warning: This is an autogenerated file.
 *
 * Changes in this file won't take effect and will be overwritten
 */

// Base Types
type Exact<T extends { [key: string]: unknown }> = { [K in keyof T]: T[K] };
type Maybe<T> = T | null | undefined
type Scalars = {
  Boolean: boolean
  String: string
  Float: number
  Int: number
  ID: string
}

// Operation related types
export type CollectionSeoFragment_StoreCollectionFragment = { seo: { title: string, description: string } };


// Query Related Code